a2a-client/requirements.txt
a2a-sdk==0.3.10